requirements.txt
django==1.11
Pillow==5.4.1